Feature refactor to only delete and add quick query

+++ resources/query.py
@@ -8,39 +8,9 @@ from schema.query import QuerySchemaBase, QueryResponseSchema
 from . import logger, requires_authorization
 
 
-class QueriesGetResource(Resource):
+class QuickQueryResource(Resource):
     @requires_authorization
-    def get(self, type):
-        """
-        ---
-        summary: Get objects queries
-        description: Returns all saved queries attached to the type of objects.
-        tags:
-            - query
-        parameters:
-            - in: path
-              name: type
-              schema:
-                type: string
-                enum: [file, config, blob, object]
-              description: Type of objects view.
-        responses:
-            200:
-                description: List of saved queries attached to the type of objects view.
-                content:
-                  application/json:
-                    schema:
-                      type: array
-                      items: QueryResponseSchema
-        """
-        quick_queries = db.session.query(Query).filter(Query.type == type).all()
-        schema = QueryResponseSchema(many=True)
-        return schema.dump(quick_queries)
-
-
-class QueryResource(Resource):
-    @requires_authorization
-    def post(self):
+    def post(self, type):
         """
         ---
         summary: Create a new query
@@ -81,43 +51,36 @@ class QueryResource(Resource):
         schema = QueryResponseSchema()
         return schema.dump(quick_query)
 
-
-class QueryUpdateResource(Resource):
     @requires_authorization
-    def put(self, id):
+    def get(self, type):
         """
         ---
-        summary: Update query.
-        description: Update custom quick query.
+        summary: Get objects queries
+        description: Returns all saved queries attached to the type of objects.
         tags:
             - query
         parameters:
             - in: path
-              name: id
+              name: type
               schema:
-                id: int
-              description: Query identifier
-        requestBody:
-            description: Basic information of query
-            content:
-              application/json:
-                schema: QuerySchemaBase
+                type: string
+                enum: [file, config, blob, object]
+              description: Type of objects view.
         responses:
             200:
-                description: When query was successfully updated
-            400:
-                description: When query is invalid
+                description: List of saved queries attached to the type of objects view.
+                content:
+                  application/json:
+                    schema:
+                      type: array
+                      items: QueryResponseSchema
         """
-        schema = QuerySchemaBase()
-        obj = schema.loads(request.get_data(as_text=True))
-        if obj.errors:
-            return {"errors": obj.errors}, 400
-        quick_query = db.session.query(Query).filter(Query.id == id).first()
-        if quick_query is not None:
-            quick_query.query = obj.data['query']
-            quick_query.name = obj.data['name']
-            db.session.commit()
+        quick_queries = db.session.query(Query).filter(Query.type == type).filter(Query.user_id == g.auth_user.id).all()
+        schema = QueryResponseSchema(many=True)
+        return schema.dump(quick_queries)
 
+
+class QuickQueryItemResource(Resource):
     @requires_authorization
     def delete(self, id):
         """
@@ -133,11 +96,14 @@ class QueryUpdateResource(Resource):
                 id: int
               description: Query identifier
         responses:
+            404:
+                descriptionL When query was not found.
             200:
                 description: When query was successfully deleted
         """
-        quick_query = db.session.query(Query).filter(Query.id == id).first()
-
+        quick_query = db.session.query(Query).filter(Query.id == id).filter(Query.user_id == g.auth_user.id).first()
+        if quick_query is None:
+            raise NotFound("Query was not found")
         if quick_query is not None:
             db.session.delete(quick_query)
             logger.info('query deleted', extra={'query': id})
